+++ lib/Storage.ts
@@ -116,17 +116,22 @@ export class LocalStorageService implements StorageService {
 /**
  * @group Storage Service
  * @description Request data interface sent by the {@link ServerSideStorageService} to get the current grid state.
+ * @typeParam TUserId The type of the user identifier.
  */
-export interface GetGridStateRequest {
-  userId: string | number
+export interface GetGridStateRequest<TUserId> {
+  /** The unique identifier for the current user.  */
+  userId: TUserId
 }
 
 /**
  * @group Storage Service
  * @description Request data interface sent by the {@link ServerSideStorageService} to save the current grid state.
+ * @typeParam TUserId The type of the user identifier.
  */
-export interface SetGridStateRequest {
-  userId: string | number
+export interface SetGridStateRequest<TUserId> {
+  /** The unique identifier for the current user.  */
+  userId: TUserId
+  /** The current grid state to save. */
   gridState: GridState
 }
 
@@ -135,8 +140,9 @@ export interface SetGridStateRequest {
  * @description Callback type to change the {@link https://developer.mozilla.org/docs/Web/API/Request | Request}
  * object before it is sent to the server from the built-in server side storage service. This is useful
  * when you need to map the {@link GetGridStateRequest} to a different data contract or alter the HTTP verb/headers.
+ * @typeParam TUserId The type of the user identifier.
  */
-export type BeforeGetRequestHandler = (request: Request, body: GetGridStateRequest) => Promise<Request>
+export type BeforeGetRequestHandler<TUserId> = (request: Request, body: GetGridStateRequest<TUserId>) => Promise<Request>
 
 /**
  * @group Storage Service
@@ -151,8 +157,9 @@ export type GetResponseHandler = (response: Response) => Promise<GridState>
  * @description Callback type to change the {@link https://developer.mozilla.org/docs/Web/API/Request | Request}
  * object before it is sent to the server from the built-in server side storage service. This is useful
  * when you need to map the {@link SetGridStateRequest} to a different data contract or alter the HTTP verb/headers.
+ * @typeParam TUserId The type of the user identifier.
  */
-export type BeforeSetRequestHandler = (request: Request, body: SetGridStateRequest) => Promise<Request>
+export type BeforeSetRequestHandler<TUserId> = (request: Request, body: SetGridStateRequest<TUserId>) => Promise<Request>
 
 /**
  * @group Storage Service
@@ -164,9 +171,12 @@ export type SetResponseHandler = (response: Response) => Promise<boolean>
 /**
  * @group Storage Service
  * @description Options to configure the built-in server-side storage service.
+ * The server-side storage service will only attempt to deserialize the response body for `getGridState`
+ * if the HTTP status code is `200 OK` and the `Content-Type` response header is `application/json`.
+ * @typeParam TUserId The type of the user identifier.
  * @see {@link ServerSideStorageService}
  */
-export interface ServerSideStorageServiceOptions {
+export interface ServerSideStorageServiceOptions<TUserId> {
   /**
    * The unique identifier for the current user that will be sent to the server with the get and set requests.
    */
@@ -181,7 +191,7 @@ export interface ServerSideStorageServiceOptions {
    * object before it is sent to the server from the built-in server side storage service. This is useful
    * when you need to map the {@link GetGridStateRequest} to a different data contract or alter the HTTP verb/headers.
    */
-  beforeGetRequest?: BeforeGetRequestHandler
+  beforeGetRequest?: BeforeGetRequestHandler<TUserId>
   /**
    * Optional callback to change the {@link https://developer.mozilla.org/docs/Web/API/Response | Response}
    * object before it is handled by the data grid. This is useful when you need to map the servers response
@@ -198,7 +208,7 @@ export interface ServerSideStorageServiceOptions {
    * object before it is sent to the server from the built-in server side storage service. This is useful
    * when you need to map the {@link SetGridStateRequest} to a different data contract or alter the HTTP verb/headers.
    */
-  beforeSetRequest?: BeforeGetRequestHandler
+  beforeSetRequest?: BeforeGetRequestHandler<TUserId>
   /**
    * Optional callback type to change the {@link https://developer.mozilla.org/docs/Web/API/Response | Response}
    * object before it is handled by the data grid from the built-in server side data service.
@@ -221,18 +231,20 @@ function getDefaultRequestOptions(): RequestInit {
 /**
  * @group Storage Service
  * @description The server-side {@link StorageService} used when {@link DataGridVueGrid.serverSideStorageOptions} is specified.
+ * This storage service will only attempt to deserialize the response body for `getGridState`
+ * if the HTTP status code is `200 OK` and the `Content-Type` response header is `application/json`.
  */
-export class ServerSideStorageService implements StorageService {
-  options: ServerSideStorageServiceOptions
+export class ServerSideStorageService<TUserId> implements StorageService {
+  options: ServerSideStorageServiceOptions<TUserId>
 
-  constructor(options: ServerSideStorageServiceOptions) {
+  constructor(options: ServerSideStorageServiceOptions<TUserId>) {
     this.options = options
   }
 
   async getGridState(): Promise<GridState | undefined> {
     const body = {
       userId: this.options.userId,
-    } as GetGridStateRequest
+    } as GetGridStateRequest<TUserId>
 
     let request = new Request(this.options.getPostRoute ?? '', {
       ...getDefaultRequestOptions(),
@@ -260,14 +272,19 @@ export class ServerSideStorageService implements StorageService {
       console.error('Failed to retrieve grid state', await response.text(), response)
       return undefined
     }
-    return response.json()
+
+    if (response.status == 200 && response.headers.get('Content-Type')?.includes('application/json')) {
+      return response.json()
+    }
+
+    return undefined
   }
 
   async setGridState(gridState: GridState): Promise<void> {
     const body = {
       userId: this.options.userId,
       gridState,
-    } as SetGridStateRequest
+    } as SetGridStateRequest<TUserId>
 
     let request = new Request(this.options.setPostRoute ?? '', {
       ...getDefaultRequestOptions(),
@@ -296,6 +313,5 @@ export class ServerSideStorageService implements StorageService {
       console.error('Failed to set grid state', await response.text(), response)
       return undefined
     }
-    return response.json()
   }
 }